Sat, Jul 29, 2023, 10:13 PM -03:00

+++ scrabble/scrabble.c
@@ -42,7 +42,7 @@ int compute_score(string word)
     {
 
 
-            int cripto = (word[i]) - 'a';
+            int cripto = ((int)word[i]) - 'a';
 
 
                 score += POINTS[cripto];
